updated homestead with php 8.1

+++ scripts/features/mongodb.sh
@@ -72,14 +72,27 @@ then
     sudo service php8.1-fpm restart
 fi
 
-phpize8.2
-./configure --with-php-config=/usr/bin/php-config8.2 > /dev/null
+if [ -f /home/$WSL_USER_NAME/.homestead-features/php81 ]
+then
+    phpize8.2
+    ./configure --with-php-config=/usr/bin/php-config8.2 > /dev/null
+    make clean > /dev/null
+    make >/dev/null 2>&1
+    sudo make install
+    sudo bash -c "echo 'extension=mongodb.so' > /etc/php/8.2/mods-available/mongo.ini"
+    sudo ln -s /etc/php/8.2/mods-available/mongo.ini /etc/php/8.2/cli/conf.d/20-mongo.ini
+    sudo ln -s /etc/php/8.2/mods-available/mongo.ini /etc/php/8.2/fpm/conf.d/20-mongo.ini
+    sudo service php8.2-fpm restart
+fi
+
+phpize"$1"
+./configure --with-php-config=/usr/bin/php-config"$1" > /dev/null
 make clean > /dev/null
 make >/dev/null 2>&1
 sudo make install
-sudo bash -c "echo 'extension=mongodb.so' > /etc/php/8.2/mods-available/mongo.ini"
-sudo ln -s /etc/php/8.2/mods-available/mongo.ini /etc/php/8.2/cli/conf.d/20-mongo.ini
-sudo ln -s /etc/php/8.2/mods-available/mongo.ini /etc/php/8.2/fpm/conf.d/20-mongo.ini
-sudo service php8.2-fpm restart
+sudo bash -c "echo 'extension=mongodb.so' > /etc/php/$1/mods-available/mongo.ini"
+sudo ln -s /etc/php/"$1"/mods-available/mongo.ini /etc/php/"$1"/cli/conf.d/20-mongo.ini
+sudo ln -s /etc/php/"$1"/mods-available/mongo.ini /etc/php/"$1"/fpm/conf.d/20-mongo.ini
+sudo service php"$1"-fpm restart
 
 mongosh admin --eval "db.createUser({user:'homestead',pwd:'secret',roles:['root']})"
